% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_cpg}
\alias{classify_cpg}
\title{Classify CpG Sites
Determines if a CpG site follows a linear, non-linear, or non-correlated trajectory.}
\usage{
classify_cpg(beta_values, age, ages_grid)
}
\arguments{
\item{beta_values}{A numeric vector of methylation beta values for a CpG site.}

\item{age}{A numeric vector of ages corresponding to the samples.}

\item{ages_grid}{A numeric vector of ages for GAM predictions.}
}
\value{
A list with classification results and GAM predictions.
}
\description{
Classify CpG Sites
Determines if a CpG site follows a linear, non-linear, or non-correlated trajectory.
}
