% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{fun}
\alias{fun}
\title{Snitch on functions}
\usage{
fun(what, col = "pink", where = topenv(parent.frame()))
}
\arguments{
\item{what}{Function to snitch on}

\item{col}{color}

\item{where}{where to look for the function to be snitched on}
}
\description{
Snitch on functions
}
\examples{
fun(rnorm)

}
