% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{pkg}
\alias{pkg}
\title{Snitch on functions from a package}
\usage{
pkg(pkg, pattern = "", col = "pink")
}
\arguments{
\item{pkg}{The package to snitch on}

\item{pattern}{filter functions}

\item{col}{color}
}
\description{
Snitch on functions from a package
}
