% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_methylation_data}
\alias{simulate_methylation_data}
\title{Generate Simulated Methylation Data}
\usage{
simulate_methylation_data(
  n_people = 300,
  seed = 123,
  plot = FALSE,
  output_dir = "./Results"
)
}
\arguments{
\item{n_people}{Number of individuals to simulate trajectories for.}

\item{seed}{Random seed for reproducibility.}

\item{plot}{TRUE/FALSE. Visualise the representative functions and the simulated data (default = FALSE).}
}
\value{
A list containing simulated methylation intensities, age data, and group labels.
}
\description{
Creates a dataset simulating DNA methylation intensity across different ageing-related patterns.
}
