% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{pkg}
\alias{pkg}
\title{Peek all functions from a package}
\usage{
pkg(pkg, pattern = "")
}
\arguments{
\item{pkg}{The package to snitch on}
}
\description{
Peek all functions from a package
}
\examples{
pkg("dplyr")
}
